ch09/ch09-api-auth-token/requirements.txt
alembic>=1.13.1
amqp>=5.2.0
asgiref>=3.7.2
async-timeout>=4.0.3
asyncpg>=0.29.0
billiard>=4.2.0
blinker>=1.7.0
celery>=5.3.6
cffi>=1.16.0
click>=8.1.7
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.3.0
colorama>=0.4.6
cryptography>=42.0.2
Flask>=3.0.2
Flask-HTTPAuth>=4.8.0
Flask-Migrate>=4.0.5
Flask-SQLAlchemy>=3.1.1
greenlet>=3.0.3
itsdangerous>=2.1.2
Jinja2>=3.1.3
kombu>=5.3.5
Mako>=1.3.2
MarkupSafe>=2.1.5
prompt-toolkit>=3.0.43
psycopg2>=2.9.9
pyclean>=3.0.0
pycparser>=2.21
PyJWT>=2.8.0
python-dateutil>=2.8.2
redis>=5.0.1
six>=1.16.0
SQLAlchemy>=2.0.25
toml>=0.10.2
typing_extensions>=4.9.0
tzdata>=2023.4
vine>=5.1.0
wcwidth>=0.2.13
Werkzeug>=3.0.1
